+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 
 * add new `contributionTypes` property to `properties` parameter ([#136])
 * add new `/contributions/latest/count` endpoint ([#174])
+* add new `/contributions/latest/count/density` endpoint
 * add new `contributionType` parameter for contributions aggregation ([#174])
 * contribution extraction endpoints now return the `@contributionChangesetId` as a separate field ([#200])